Shut up codefactor.io
